+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Current
 - Allows to specify a custom mount path on namespace registration
 - Allow to express models as raw schemas
 - Upgraded to Swagger-UI 2.2.6
+- Support Swagger-UI translations
 
 0.9.2 (2016-04-22)
 ------------------